refactor: remove deprecated root store file and update import paths across the application
