--- apps/web/ce/store/root.store.ts
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) 2023-present Plane Software, Inc. and contributors
- * SPDX-License-Identifier: AGPL-3.0-only
- * See the LICENSE file for details.
- */
-
-// store
-import { CoreRootStore } from "@/store/root.store";
-import type { ITimelineStore } from "@/store/timeline/timeline.store";
-import { TimeLineStore } from "@/store/timeline/timeline.store";
-
-export class RootStore extends CoreRootStore {
-  timelineStore: ITimelineStore;
-
-  constructor() {
-    super();
-
-    this.timelineStore = new TimeLineStore(this);
-  }
-}
