+++ apps/web/core/lib/store-context.tsx
@@ -7,7 +7,7 @@
 import type { ReactElement } from "react";
 import { createContext } from "react";
 // plane web store
-import { RootStore } from "@/plane-web/store/root.store";
+import { RootStore } from "@/store/root.store";
 
 export let rootStore = new RootStore();
