+++ apps/web/core/store/member/index.ts
@@ -11,7 +11,7 @@ import type { IUserLite } from "@plane/types";
 // plane web imports
 import type { IProjectMemberStore } from "@/store/member/project/base-project-member.store";
 import { ProjectMemberStore } from "@/store/member/project/base-project-member.store";
-import type { RootStore } from "@/plane-web/store/root.store";
+import type { RootStore } from "@/store/root.store";
 // local imports
 import type { IWorkspaceMemberStore } from "./workspace/workspace-member.store";
 import { WorkspaceMemberStore } from "./workspace/workspace-member.store";
