+++ apps/web/core/store/pages/project-page.store.ts
@@ -15,7 +15,7 @@ import { EUserProjectRoles } from "@plane/types";
 import { filterPagesByPageType, getPageName, orderPages, shouldFilterPage } from "@plane/utils";
 // plane web constants
 // plane web store
-import type { RootStore } from "@/plane-web/store/root.store";
+import type { RootStore } from "@/store/root.store";
 // services
 import { ProjectPageService } from "@/services/page";
 // store
