+++ apps/web/core/store/pages/project-page.ts
@@ -10,7 +10,7 @@ import { computedFn } from "mobx-utils";
 import { EPageAccess, EUserPermissions } from "@plane/constants";
 import type { TPage } from "@plane/types";
 // plane web store
-import type { RootStore } from "@/plane-web/store/root.store";
+import type { RootStore } from "@/store/root.store";
 // services
 import { ProjectPageService } from "@/services/page";
 const projectPageService = new ProjectPageService();
